Timer works in milliseconds, xml parser improved to parse stations, lines and trams in another way. Tracks will be implicitly parsed

+++ src/utils/constants.h
@@ -26,7 +26,7 @@ namespace constants {
      * The time the program should run
      */
 
-    extern const int RUNTIME_SECONDS;
+    extern const double RUNTIME_SECONDS;
 
     /**
      *
